% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-level.R
\name{power1L}
\alias{power1L}
\title{power1L}
\usage{
power1L(n, delta, numcovL1 = 0, R2L1 = 0, alpha = 0.05)
}
\arguments{
\item{n}{Number of subjects in level 1 (individual level).}

\item{delta}{Standardized effect size.}

\item{numcovL1}{Number of covariate(s) at level 1. Default value is \code{0}.}

\item{R2L1}{R-squared at level 1. Default value is \code{0}.}

\item{alpha}{Probability of type I error. Default value is \code{0.05}.}
}
\description{
This function returns the power for single-level randomized trials designs. It addresses the following models:
\itemize{
\item Main treatment effects, with no additional covariates.
\item Main treatment effects, with covariate(s) at level 1.
}
}
\examples{
# All arguments are specified
power1L(n=100,delta=0.2,numcovL1=1,R2L1=0.2,alpha=0.05)

# Unspecified arguments take their default values
power1L(n=100,delta=0.2)
}
\seealso{
\code{\link{PowerUpR}}
}
