% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite-CRT.R
\name{powerMSCRT}
\alias{powerMSCRT}
\title{powerMSCRT}
\usage{
powerMSCRT(n, J, K, delta, sigmad, rhoL2, rhoL3, numcovL1 = 0, numcovL2 = 0,
  R2L1 = 0, R2L2 = 0, randeff = TRUE, alpha = 0.05)
}
\arguments{
\item{n}{Number of subjects in level 1 (individual level).}

\item{J}{Number of units in level 2 (cluster level).}

\item{K}{Number of units in level 3 (site level).}

\item{delta}{Standardized effect size.}

\item{sigmad}{Variance in treatment effects across sites.}

\item{rhoL2}{Intraclass correlation (ICC) at level 2.}

\item{rhoL3}{Intraclass correlation (ICC) at level 3.}

\item{numcovL1}{Number of covariate(s) at level 1. Default value is \code{0}.}

\item{numcovL2}{Number of covariate(s) at level 2. Default value is \code{0}.}

\item{R2L1}{R-squared at level 1. Default value is \code{0}.}

\item{R2L2}{R-squared at level 2. Default value is \code{0}.}

\item{randeff}{Random site effects. Default value is \code{FALSE}.}

\item{alpha}{Probability of type I error. Default value is \code{0.05}.}
}
\description{
This function returns the power for multi-site cluster randomized trials designs with treatment at level 2 (cluster level). It addresses the following models:
\itemize{
\item Main treatment effects (assuming random site effects) with no additional covariates.
\item Main treatment effects (assuming random site effects) with covariate(s) at level 1 only.
\item Main treatment effects (assuming random site effects) with covariate(s) at level 2 only.
\item Main treatment effects (assuming random site effects) with covariate(s) at levels 1 and 2.
\item Main treatment effects (assuming fixed site effects) with no additional covariates.
\item Main treatment effects (assuming fixed site effects) with covariate(s) at level 1 only.
\item Main treatment effects (assuming fixed site effects) with covariate(s) at level 2 only.
\item Main treatment effects (assuming fixed site effects) with covariate(s) at levels 1 and 2.
}
}
\examples{
# All arguments are specified
powerMSCRT(n=100,J=40,K=40,delta=0.2,sigmad=0.2,rhoL2=0.3,rhoL3=0.2,numcovL1=1,numcovL2=1,R2L1=0.2,R2L2=0.30,randeff=TRUE,alpha=0.05)

# Unspecified arguments take their default values
powerMSCRT(n=100,J=40,K=40,delta=0.2,sigmad=0.2,rhoL2=0.3,rhoL3=0.2)
}
\seealso{
\code{\link{PowerUpR}}
}
